Split movement from positioner to mover
