+++ luda-adventure/src/app/game/types/game_object/types/mod.rs
@@ -1,7 +1,9 @@
 mod collider;
+mod mover;
 mod positioner;
 mod renderer;
 
 pub use collider::*;
+pub use mover::*;
 pub use positioner::*;
 pub use renderer::*;
