+++ luda-adventure/src/app/game/update/evaluate_ticks/move_character.rs
@@ -1,13 +1,17 @@
-use crate::app::game::{Game, PlayerCharacter, Positioner, TICK_INTERVAL};
+use crate::app::game::{Game, Movement, Mover, PlayerCharacter, Positioner, TICK_INTERVAL};
+use namui::prelude::*;
 
 impl Game {
     pub fn move_character(&mut self) {
-        if let Some((_entity, (_player_character, positioner))) = self
+        if let Some((_entity, (_player_character, positioner, mover))) = self
             .ecs_app
-            .query_entities_mut::<(&mut PlayerCharacter, &mut Positioner)>()
+            .query_entities_mut::<(&PlayerCharacter, &mut Positioner, &Mover)>()
             .first_mut()
         {
-            positioner.apply_movement(TICK_INTERVAL);
+            if let Movement::Moving(velocity) = mover.movement() {
+                let delta_xy = velocity * Xy::single(TICK_INTERVAL);
+                positioner.set_xy(positioner.xy() + delta_xy);
+            }
         }
     }
 }
